+++ src/lib/constants/sections.ts
@@ -255,6 +255,13 @@ export const SCHEMA_SECTIONS = {
         titlePlural: i18n.t('Constants'),
         parentSectionKey: 'other',
     },
+    optionGroup: {
+        name: SchemaName.optionGroup,
+        namePlural: 'optionGroups',
+        title: i18n.t('Option group'),
+        titlePlural: i18n.t('Option groups'),
+        parentSectionKey: 'other',
+    },
     attribute: {
         name: SchemaName.attribute,
         namePlural: 'attributes',