+++ src/pages/dataSetsWip/form/dataEntryForm/SectionFormList.tsx
@@ -1,6 +1,6 @@
 import i18n from '@dhis2/d2-i18n'
-import { Button, IconAdd16 } from '@dhis2/ui'
-import React, { useState } from 'react'
+import { Button, ButtonStrip, IconAdd16 } from '@dhis2/ui'
+import React, { ComponentProps, useState } from 'react'
 import { useFieldArray } from 'react-final-form-arrays'
 import {
     StandardFormSectionTitle,
@@ -15,13 +15,15 @@ import { DisplayableModel } from '../../../../types/models'
 import { DataSetValues } from '../../Edit'
 import { EditOrNewDataSetSectionForm } from './sectionForm/DataSetSectionForm'
 import css from './SectionFormList.module.css'
+import { SectionsOrderingModal } from './SectionsOrderingModal'
 
 type Section = DataSetValues['sections'][number]
 
 export const SectionFormSectionsList = () => {
-    const [sectionFormOpen, setSectionFormOpen] = React.useState<
+    const [sectionFormOpen, setSectionFormOpen] = useState<
         DisplayableModel | null | undefined
     >(undefined)
+    const [orderSectionsFormOpen, setOrderSectionsFormOpen] = useState(false)
     // use null as open, but new model
     const isSectionFormOpen = !!sectionFormOpen || sectionFormOpen === null
     const sectionFieldArray = useFieldArray<Section>('sections').fields
@@ -40,7 +42,7 @@ export const SectionFormSectionsList = () => {
         })
     }
 
-    const handleSubmittedSection: React.ComponentProps<
+    const handleSubmittedSection: ComponentProps<
         typeof EditOrNewDataSetSectionForm
     >['onSubmitted'] = (values) => {
         const isEditSection = sectionFormOpen && sectionFormOpen.id
@@ -71,6 +73,13 @@ export const SectionFormSectionsList = () => {
                     />
                 )}
             </DrawerPortal>
+            {orderSectionsFormOpen && (
+                <SectionsOrderingModal
+                    onClose={() => setOrderSectionsFormOpen(false)}
+                    sections={sectionFieldArray.value}
+                    onReorder={sectionFieldArray.update}
+                />
+            )}
             <div>
                 <StandardFormSectionTitle>
                     {i18n.t('Sections')}
@@ -116,14 +125,24 @@ export const SectionFormSectionsList = () => {
                 })}
             </div>
             <div>
-                <Button
-                    secondary
-                    small
-                    icon={<IconAdd16 />}
-                    onClick={() => setSectionFormOpen(null)}
-                >
-                    {i18n.t('Add section')}
-                </Button>
+                <ButtonStrip>
+                    <Button
+                        secondary
+                        small
+                        icon={<IconAdd16 />}
+                        onClick={() => setSectionFormOpen(null)}
+                    >
+                        {i18n.t('Add section')}
+                    </Button>
+                    <Button
+                        secondary
+                        small
+                        onClick={() => setOrderSectionsFormOpen(true)}
+                        disabled={sectionFieldArray.value.length <= 1}
+                    >
+                        {i18n.t('Reorder sections')}
+                    </Button>
+                </ButtonStrip>
             </div>
         </div>
     )